requirements.txt
asgiref==3.9.1
django-cors-headers==4.7.0
stripe==11.4.1
drf-spectacular==0.28.0
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
cryptography==45.0.5
defusedxml==0.7.1
dj-database-url==3.1.2
dj-rest-auth==4.0.0
Django==4.2.23
django-allauth==0.52.0
djangorestframework==3.16.0
dotenv==0.9.9
idna==3.10
oauthlib==3.3.1
psycopg2-binary
pycparser==2.22
PyJWT==2.10.1
python-dotenv==1.1.1
python3-openid==3.2.0
requests==2.32.4
requests-oauthlib==2.0.0
sqlparse==0.5.3
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
